Move bom_types to namespace submodule
