+++ doc/source/api/bom_builder/api.rst
@@ -11,8 +11,8 @@ see :ref:`ref_grantami_bomanalytics_api_index`.
 .. note::
    The following elements are defined in the 23/01 BoM schema but are not supported in this module:
 
-   * ``NonMIPartReference`` on :class:`~ansys.grantami.bomanalytics.bom_types._bom_types.Part`
-   * ``Annotations`` and ``AnnotationSources`` on :class:`~ansys.grantami.bomanalytics.bom_types._bom_types.BillOfMaterials`
+   * ``NonMIPartReference`` on :class:`~ansys.grantami.bomanalytics.bom_types.eco2301._bom_types.Part`
+   * ``Annotations`` and ``AnnotationSources`` on :class:`~ansys.grantami.bomanalytics.bom_types.eco2301._bom_types.BillOfMaterials`
 
    This module does not contain classes that correspond to these types and cannot serialize a BoM that includes these
    elements. It is still possible to deserialize an XML BoM that uses these elements, but these elements cannot be
@@ -21,12 +21,12 @@ see :ref:`ref_grantami_bomanalytics_api_index`.
 .. versionadded:: 2.0
 
 .. _ref_grantami_bomanalytics_api_billofmaterials:
-.. autoclass:: ansys.grantami.bomanalytics.bom_types._bom_types.BillOfMaterials
+.. autoclass:: ansys.grantami.bomanalytics.bom_types.BillOfMaterials
    :inherited-members:
    :member-order: bysource
 
 
-.. automodule:: ansys.grantami.bomanalytics.bom_types._bom_types
-   :exclude-members: BillOfMaterials, BaseType, HasNamespace, SupportsCustomFields
+.. automodule:: ansys.grantami.bomanalytics.bom_types
+   :exclude-members: BillOfMaterials, BaseType2301, HasNamespace, SupportsCustomFields
    :inherited-members:
    :member-order: bysource
