+++ src/ansys/grantami/bomanalytics/bom_types/eco2301/__init__.py
@@ -20,6 +20,6 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from ._builders import *
-from .eco2301 import *
-from .gbt1205 import *
+from ._bom_reader import BoMReader
+from ._bom_types import *
+from ._bom_writer import BoMWriter
