+++ src/ansys/grantami/bomanalytics/bom_types/gbt1205/__init__.py
@@ -20,6 +20,4 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from ._builders import *
-from .eco2301 import *
-from .gbt1205 import *
+from ._bom_types import *
